Added Create Event functionality

+++ app/controllers/calendars_controller.rb
@@ -1,7 +1,12 @@
 class CalendarsController < ApplicationController
+  before_filter :authorize
 	def show
-		calendar = Calendar.find(params[:id])
-		render :json => calendar
+    calendar = Calendar.find(params[:id])    
+    if current_user.has_show_permission?(calendar)
+      render :json => calendar
+    else
+      render :json => "Permission denied", :status => :forbidden
+		end
 	end
 
 	def create
@@ -17,25 +22,33 @@ class CalendarsController < ApplicationController
 			end
 			render :json => calendar
 		rescue
-			render :json => calendar.errors.full_messages, :status => 422
+			render :json => calendar.errors.full_messages * ", " , :status => 422
 		end
 	end
 
 	def update
-		calendar = current_user.calendars.find(params[:id])
-		if calendar.update_attributes(params[:calendar])
-			render :json => calendar
-		else
-			render :json => calendar.errors.full_messages
-		end
+		calendar = Calendar.find(params[:id])
+		if current_user.has_update_permission?(calendar)
+			if calendar.update_attributes(params[:calendar])
+				render :json => calendar
+			else
+		    render :json => calendar.errors.full_messages * ", "
+			end
+    else
+      render :json => "Permission denied.", :status => :forbidden
+    end
 	end
 
 	def destroy
-		calendar = current_user.calendars.find(params[:id])
-		if calendar.destroy
-			render :json => calendar
-		else
-			render :json => calendar.errors.full_messages
-		end
-	end
+		calendar = Calendar.find(params[:id])
+    if current_user.has_update_permission?(calendar)
+  		if calendar.destroy
+  			render :json => calendar
+  		else
+  			render :json => calendar.errors.full_messages * ", "
+  		end
+    else
+      render :json => "Permission denied.", :status => :forbidden
+    end
+  end
 end
